Customer/requirement.txt
Django
djangorestframework
requests
drf_writable_nested